[start 6.0] +ranking month chart, +month chart upper limit

+++ src/xyz/spiralhalo/sherlock/persist/cache/CacheHandler.java
@@ -32,13 +32,14 @@ class CacheHandler {
         return null;
     }
 
-    static void forceDiskCacheCleanup() {
+    static void deleteAllCacheFiles() {
         File cacheDir = new File(Application.getCacheDir());
         if(cacheDir.exists()) {
             File[] cacheFiles = cacheDir.listFiles();
             if (cacheFiles != null) {
                 for (File file : cacheFiles) {
-                    if (!file.getName().contains(".")) {
+                    // this function only deletes files without extensions
+                    if (!file.getName().contains(".") && !file.isDirectory()) {
                         file.delete();
                     }
                 }
